Follow-up commit, cleaned version of SLASH'EM Extended v97 (11 feb 2015)
There was no version control between SLASH'EM 0.0.7F7E2 and this, sadly.

+++ include/winX.h
@@ -53,7 +53,7 @@ struct text_map_info_t {
 };
 
 struct tile_map_info_t {
-    unsigned short glyphs[ROWNO][COLNO];	/* Saved glyph numbers. */
+    unsigned /*short*/long glyphs[ROWNO][COLNO];	/* Saved glyph numbers. */
     GC	white_gc;
     GC	black_gc;
 };
